for loop error known bug
so we try to use a for loop to add all of the meshes to the scene (there are some 6 to make the mask) but we get different incomplete results each reload. This is because it is cutting and pasting from the object scene to the scene we are adding the object, and then it goes to the "next" mesh in line rather that the one at the top fo the list. Basically, it winds up ignoring 2-3 meshes each time.