animation - enabling animation clip
similar to animation keyframes, we need to create an animation mixer. Much like a player for songs.

First, we need an animated model. We can use the fox located in /static/models/Fox/ folder (also taken from the glTF model samples).
Change the path to load that fox:
We have a problem; the fox is way too big. If you can't see it, look above or zoom out.

Before handling the animation, let's fix the scale. If you look at the composition of the imported scene, the fox is composed of one Object3D, itself made of a Bone and a SkinnedMesh. We won't explain what those are, but the idea is that we shouldn't simply scale the Object3D. Even if it would work in this case, it would probably not work with more complex models.

What we can do here, is scale the loaded scene and add it directly to our scene: