Model continuted
Because it's a complex model, we will simply add the gltf.scene group to our own scene:

the helmet is hidding inside of the sphere. You can see this by zooming in.

We removed the sphere and scaled up the helmet. We also lowered -4 on the y axis to sit better in our scene, and slighting rotated the helmet for a better angle.