Environment map 1
We can't see much of our model because we have only one weak DirectionalLight. As we said previously, the lighting will be taken care of by the environment map.

We already talked about the environment map in the Materials lesson. An environment map is like a photo of the surrounding. It can be a 360 photo or 6 photos that, once put together, compose a cube.

We will use the environment map both for the background and to illuminate our model.

Load the environment map
First, let's load our environment map. There are multiple textures located in the /static/textures/environmentMaps/ folder. We are going to use the first one.

Because these textures are composed of 6 images (a cube), we have to use a CubeTextureLoader.

Add the CubeTextureLoader to our loaders:

Now we can load the textures. The order is positive x, negative x, positive y, negative y, positive z, and negative z.

Add it after creating the scene: