Apply the environment map to the model
One essential feature to get a realistic render is to use our environment map to lighten our model.

We have already covered how to apply an environment map to a MeshStandardMaterial with the envMap property. The problem is that our model is composed of many Meshes. What we can do is use the traverse(...) method available on every Object3D —and classes that inherit from it like Group and Mesh.

Instead of doing it in the success callback function, we will create a updateAllMaterials function that will get handy later. Create this function before the environment map: