change the envMap property
You should see all the children and grand children in the logs.

Instead of logging the children, we want to apply the environment map to each material that should have it.

It would make no sense to apply the environment map to the lights, the camera, the group, etc. We only want to apply the environment map to the Meshes that have a MeshStandardMaterial.

What we can do is test if the child is an instance of THREE.Mesh and if its material is an instance of THREE.MeshStandardMaterial:

We now only log the children that support environment maps. Let's change their envMap property in the material property:

Can't see much of a difference? Increase the envMapIntensity to 5: