add envMap intensity to datGUI
That's better. We get a nice and realistic lighting.

For more control, let's add the envMapIntensity property to our Dat.GUI. The problem is that we need only one property to tweak, and, when changed, this value should be applied to all children materials.

We can, however, use the debugObject technique as we did in a previous lesson. Right after instantiating Dat.GUI, create a debugObject:

Then, in the environment map code section, add an envMapIntensity property to that object as well as to your Dat.GUI:

We now have the envMapIntensity tweak, but changing the value isn't updating the scene and its children. We should now call the updateAllMaterials function when the tweak value changes and use the debugObject.envMapIntensity value in the updateAllMaterials function:

We can now change all Meshes environment map intensity with only one tweak directly in our debug interface.