environment map as default
There is an easier way of applying the environment map to all objects. We can update the environment property of the scene just like we changed the background property:

this is another option. But we need to control the intensity of every mesh envmap and that can only be done the way we did previously.