Apply the environment map as default
There is an easier way of applying the environment map to all objects. We can update the environment property of the scene just like we changed the background property:

Unluckily, we cannot change the environment map intensity of each material directly from the scene, so we still need our updateAllMaterials function, but it's a perfectly viable solution.