Renderer - Output encoding
Things are getting more and more realistic, but we still feel that it's all fake. We need to work on the colors, and this is a matter of WebGLRenderer properties.

Without going too much into details, the outputEncoding property controls the output render encoding. By default, the value of outputEncoding is THREE.LinearEncoding, which looks ok, but not realistic.

The recommended value for the outputEncoding is THREE.sRGBEncoding:
renderer.outputEncoding = THREE.sRGBEncoding

You should see much brighter textures that will also impact the environment map.