another option: gamma encoding
Another possible value is THREE.GammaEncoding. This encoding has the advantage of letting you play on a value called gammaFactor that would act a little like the brightness, but we won't use this one in the lesson.

The Gamma Encoding is a way of storing colors while optimizing how bright and dark values are stored according to human eye sensitivity. When we use the sRGBEncoding, it's like using the GammaEncoding with a default gamma factor of 2.2, which is the common value.

You can find out more about this topic here

https://www.donmccurdy.com/2020/06/17/color-management-in-threejs/
https://medium.com/game-dev-daily/the-srgb-learning-curve-773b7f68cf7a
While some might think that GammaEncoding is better than sRGBEncoding because we can control the gamma factor for a darker or brighter scene, this physically doesn't seem right, and we will see how to manage the "brightness" in a better way later.