Textures encoding
You might not have noticed it, but the environment map colors are wrong. They appear grayish and toned down. Even if the effect looks pretty good, it's more satisfying to preserve the right colors.

The problem is that our renderer outputEncoding is THREE.sRGBEncoding, yet the environment map texture is by default THREE.LinearEncoding.

The rule is straightforward. All textures that we can see directly —like the map —should have THREE.sRGBEncoding as encoding, and all other textures — such as normalMap —should have THREE.LinearEncoding.

We can see the environmentMap texture directly, so we have to change its encoding to THREE.sRGBEncoding:

But what about the model textures?

Fortunately, the GLTFLoader implemented this rule, and all the textures loaded from it will have the right encoding automatically.