add toneMapping to dat.gui
To appreciate the difference, let's add the toneMapping to our Dat.GUI. We can create a dropdown tweak by sending an object with different keys and values as the third parameter of gui.add(...):

Unfortunately, changing this tweak will result in a warning in the console, looking likeTHREE.WebGLProgram: Unsupported toneMapping: 3. The problem is a bug with Dat.GUI interpreting the values of the object as a String.

For example, if we use THREE.ReinhardToneMapping, the value behind this constant is 3 as a Number. But once we change the tweak for tone mapping, Dat.GUI will send '3' as a String which will result in the error previously mentioned.

We can fix that by converting the value to a Number in the change event callback:

Changing the tone mapping should work. Nevertheless, if you watch closely, you'll see that the tone mapping changed for the environment map in the background, but not for the model itself.