fixing dat.gui toneMapping on model
Changing the tone mapping should work. Nevertheless, if you watch closely, you'll see that the tone mapping changed for the environment map in the background, but not for the model itself.

We need to find a way to tell the materials that they need to be updated and we already did that in the updateAllMaterials function. What we can do is call this function right after changing the toneMapping:

The materials should also update when changing the tone mapping.