Shadows + using helper to align camera shadow
The final touch for a realistic render is to add shadows. First, toggle the shadows on WebGLRenderer. Then, change the shadow type to THREE.PCFSoftShadowMap as we did in the Shadows lesson:

renderer.shadowMap.enabled = true
renderer.shadowMap.type = THREE.PCFSoftShadowMap

Activate it on the DirectionalLight:

We also need to set the camera that handles the shadow for this light.

Add a CameraHelper to the directionalLight.shadow.camera:

We can now see accurately what the shadow camera will render. The box should already fit pretty nicely with the scene. Let's reduce the far value: