shadows cont
We can remove or comment the directionalLightCameraHelper.

As we want realistic and precise shadows and because we have only one light, we can increase the shadow map size to 1024x1024 without fearing a frame rate drop.

Finally, we can activate the shadows on all the Meshes of our model. As we are already traversing the scene in the updateAllMaterials function, let's simply activate both castShadow and receiveShadow on all the children:

set castShadow and receiveShadow to true