add burger - see shadow acne
Let's try with our hamburger. A version is already located in /static/models/hamburger.glb.

This file isn't Draco compressed. If you are using your model, make sure it's not compressed or add the DRACOLoader to the GLTFLoader as we did in the Imported Model lesson.

Replace the path to load the hamburger and change the scale and position:

Your hamburger appears, but some nasty strips cover its surface.

No we didn't let the hamburger burn on the grill.

These artifacts are called shadow acne. Shadow acne can occur on both smooth and flat surfaces for precision reasons when calculating if the surface is in the shadow or not. What's happening here is that the hamburger is casting a shadow on its own surface.

We have to tweak the light shadow's bias and normalBias properties to fix this shadow acne.

The bias usually helps for flat surfaces. It's not our case here, but if you have the problem on flat surfaces, try to increase the bias slightly until the acne disappears.

The normalBias usually helps for rounded surfaces, which is our case. Let's increase it until the shadow acne is barely visible:

I also added a datGUI tweak to see it in action :D