Import / update webpack
Unfortunately, we get a Webpack error. We need to tell Webpack how to handle .glsl files.

To do this, go to the /bundler/webpack.common.js where loaders are set in the rules array property.

Add the following rule at the rules —or anywhere, as long as it's in the array —and don't forget the comma between those rules:

This rule solely tells Webpack to provide the raw content of the files having .glsl, .vs, .fs, .vert or .frag as extension. We might as well have used .glsl, but some people feel more comfortable using other file extensions.

Re-launch the server with npm run dev and the Webpack error will disappear.