Vector 3
vec3 is just like vec2, but with a third property named z. It's very convenient when one needs 3D coordinates:
vec3 foo = vec3(0.0);
vec3 bar = vec3(1.0, 2.0, 3.0);
bar.z = 4.0;

While we can use x, y, and z, we can also work with r, g, and b. This is just syntax sugar and the result is exactly the same. It's very effective when we use a vec3 to store colors:

vec3 purpleColor = vec3(0.0);
purpleColor.r = 0.5;
purpleColor.b = 1.0;
GLSL
A vec3 can also be partially created from a vec2:

vec2 foo = vec2(1.0, 2.0);
vec3 bar = vec3(foo, 3.0);
GLSL
We can also take a part of the vec3 to generate a vec2:

vec3 foo = vec3(1.0, 2.0, 3.0);
vec2 bar = foo.xy;
GLSL
Here, the bar will be a vec2 with 1.0, and 2.0 as values.

This is called a swizzle and we can also use the properties in a different order:

vec3 foo = vec3(1.0, 2.0, 3.0);
vec2 bar = foo.yx;