Functions - until 1:02:00
Just like in most programming languages, we can create and use functions.

A function must start with the type of the value that will be returned:

float loremIpsum()
{
    float a = 1.0;
    float b = 2.0;

    return a + b;
}
GLSL
If the function isn't supposed to return anything, we set the type to void:

void justDoingStuff()
{
    float a = 1.0;
    float b = 2.0;
}
GLSL
We can specify parameters, but we also have to provide their type:

float add(float a, float b)
{
    return a + b;
}
GLSL
This function is, as you can imagine, worthless.