Position attributes
First, we retrieve the vertex position with:

attribute vec3 position;
GLSL
Remember that the same code applies to every vertices of the geometry. Attributes are the only variable that will change between the vertices. The same vertex shader will be applied for each vertex and the position attribute will contain the x, y, and z coordinates of that specific vertex.

Then, we convert this vec3 to a vec4:

gl_Position = /* ... */ vec4(position, 1.0);
GLSL
This is because the following matrices and the gl_Position need to use vec4 as we saw earlier.