Matrices uniforms cont
There are 3 matrices in our code, and because their values are the same for all the vertices of the geometry, we retrieve them by using uniforms.

uniform mat4 projectionMatrix;
uniform mat4 viewMatrix;
uniform mat4 modelMatrix;
GLSL
Each matrix will do a part of the transformation:

The modelMatrix will apply all transformations relative to the Mesh. If we scale, rotate or move the Mesh, these transformations will be contained in the modelMatrix and applied to the position.
The viewMatrix will apply transformations relative to the camera. If we rotate the camera to the left, the vertices should be on the right. If we move the camera in direction of the Mesh, the vertices should get bigger, etc.
The projectionMatrix will finally transform our coordinates into the final clip space coordinates.
If you want to find out more about those matrices and coordinates, here's a good article: https://learnopengl.com/Getting-started/Coordinate-Systems.

To apply a matrix, we multiply it. If want to apply a mat4 to a variable, this variable has to be a vec4. We can also multiply matrices with other matrices:

gl_Position = projectionMatrix * viewMatrix * modelMatrix * vec4(position, 1.0);
GLSL
There is actually a shorter version where the viewMatrix and the modelMatrix are combined into a modelViewMatrix:

uniform mat4 projectionMatrix;
uniform mat4 modelViewMatrix;

attribute vec3 position;

void main()
{
    gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
}
GLSL
This is shorter, but we have less control over each step.

We will actually make our code even longer to have a better understanding and more control on the position:

uniform mat4 projectionMatrix;
uniform mat4 viewMatrix;
uniform mat4 modelMatrix;

attribute vec3 position;

void main()
{
    vec4 modelPosition = modelMatrix * vec4(position, 1.0);
    vec4 viewPosition = viewMatrix * modelPosition;
    vec4 projectedPosition = projectionMatrix * viewPosition;

    gl_Position = projectedPosition;
}
GLSL
Those changes have the exact same result but we can now move the whole model just by tweaking the values of the modelPosition:

void main()
{
    vec4 modelPosition = modelMatrix * vec4(position, 1.0);
    modelPosition.y += 1.0;

    // ...
}

The whole plane should appear higher.

Or we can do cooler things such as transforming our plane wave:

void main()
{
    vec4 modelPosition = modelMatrix * vec4(position, 1.0);
    modelPosition.z += sin(modelPosition.x * 10.0) * 0.1;

    // ...
}
We changed the z by using the x coordinate through a sin(...) function. Good luck getting this result with Three.js built-in materials.

it was too hard to see so we had to multiply the model position of x by 10 to increase the intensity. THen we had to multiply everything by .1 to increase frequency

I don't have a fucking clue at all.