Understanding the fragment shader (precision + fragColor)
The fragment shader code will be applied to every visible fragment of the geometry. That is why the fragment shader comes after the vertex shader.

The code is more manageable than the vertex shader.

Main function
Again, we face the main function:

void main()
{
}

Precision
We also have an instruction at the top of the code:

precision mediump float;
GLSL
This instruction lets us decide how precise can a float be. There are different possible values:

highp
mediump
lowp
highp can have performance hit and might not even work on some devices. lowp can create bugs by the lack of precision. We ordinarily use mediump. We also could have set the precision for the vertex shader but it's not required.

This part is automatically handled when we are using ShaderMaterial instead of RawShaderMaterial.

gl_FragColor
The gl_FragColor is like the gl_Position but for the color. It's already declared, and we need to assign it in the main function.

It's a vec4 with the first three values being the red, green, and blue channels (r, g, b) and the fourth value being the alpha (a):

gl_FragColor = vec4(0.5, 0.0, 1.0, 1.0);

This code will result on a purple color for the whole geometry.

Each property of gl_FragColor goes from 0.0 to 1.0. We can go beyond those values with no error, but it won't help us.

If we want to set an alpha below 1.0, we also need to set the transparent property to true in the RawShaderMaterial:

const material = new THREE.RawShaderMaterial({
    vertexShader: testVertexShader,
    fragmentShader: testFragmentShader,
    transparent: true
})